refactor: Refactoring for sqlite

+++ crates/machine/src/machine/machine.rs
@@ -29,7 +29,7 @@ pub struct MachineData {
     pub status: MachineStatus,
     pub motors_enabled: bool,
     pub gcode_history: VecDeque<GCodeHistoryEntry>,
-    pub paused_task_id: Option<u64>, 
+    pub paused_task_id: Option<u32>,
 }
 
 impl Actor for Machine {}
